Pass contact to CelebrationIcons to keep date logic isolated
- Update mock date to Jan instead of Feb
- Create CelebrationItemsFragment
